requirements.txt
SQLAlchemy==1.4.32
pg8000==1.24.1
nsj_gcf_utils==1.0.9
pydantic==1.9.0
authlib==1.0.0
ptvsd==4.3.2
nsj-flask-auth==0.3.0
nsj-rest-lib==0.1.6
python-logging-loki>=0.3.1
# Só deixar o requirement abaixo se for usar processamento assincrono
pika==1.3.1
sentry-sdk==1.5.12
PyMySQL==1.1.0
# Requirements que se destinam ao exemplo de comunicação assíncrona
elasticsearch==8.8.2
nsj-queue-lib==0.5.6
nsj_sql_utils_lib==0.0.4
psycopg2_binary>=2.9.6
nsj-queue-lib==0.5.6
